Custom code execution file and start delay.
You can now enter custom script to execute before starting an instance. This is would allow an admin to delete the ocean.atlas save file if desired before starting the instance until Grapeshot fixes the issue.
A countdown timer will show from a predetermined amount of time. You can set in the atlasmanger.cfg file how long this is or set it to 0 for no timer. If no entry is in the cfg file it will default to 30 seconds. The countdown will stop once it detects the server as online.